+++ lib/src/builders/select.dart
@@ -171,25 +171,3 @@ final class AfterFetch extends QueryBuilder
     with Build, TimeoutClause, ParallelClause {
   const AfterFetch(super._query);
 }
-
-/// Handles `TIMEOUT`
-base mixin TimeoutClause on QueryBuilder {
-  AfterTimeout timeout({
-    required String duration,
-  }) {
-    _query.addAll(['TIMEOUT', duration]);
-    return AfterTimeout(_query);
-  }
-}
-
-final class AfterTimeout extends QueryBuilder with Build, ParallelClause {
-  const AfterTimeout(super._query);
-}
-
-/// Handles `PARALLEL`
-base mixin ParallelClause on QueryBuilder {
-  End parallel() {
-    _query.add('PARALLEL');
-    return End(_query);
-  }
-}