+++ lib/surrealdb_query_builder.dart
@@ -1,11 +1,14 @@
 /// A Very Good Project created by Very Good CLI.
 library surrealdb_query_builder;
 
+import 'dart:convert';
+
 import 'package:surrealdb_query_builder/src/utils/list_utils.dart';
 
 part './src/builders/common/clauses.dart';
 part './src/builders/common/logical_op.dart';
 part './src/builders/common/op.dart';
+part './src/builders/create.dart';
 part './src/builders/live_select.dart';
 part './src/builders/select.dart';
 part './src/core/constants/enums.dart';